Apply pdfium_noshorten_config to xfa/fxfa/parser
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/90812
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ fxjs/xfa/cjx_object.cpp
@@ -1184,7 +1184,7 @@ void CJX_Object::ScriptSomBorderWidth(v8::Isolate* pIsolate,
     return;
 
   WideString wsThickness = fxv8::ReentrantToWideStringHelper(pIsolate, *pValue);
-  for (int32_t i = 0; i < border->CountEdges(); ++i) {
+  for (size_t i = 0; i < border->CountEdges(); ++i) {
     CXFA_Edge* edge = border->GetEdgeIfExists(i);
     if (edge)
       edge->SetMSThickness(CXFA_Measurement(wsThickness.AsStringView()));
