+++ xfa/fxfa/parser/cxfa_node.h
@@ -364,7 +364,7 @@ class CXFA_Node : public CXFA_Object, public GCedTreeNodeMixin<CXFA_Node> {
   CXFA_Node* GetExclGroupNextMember(CXFA_Node* pNode);
 
   bool IsChoiceListAllowTextEntry();
-  int32_t CountChoiceListItems(bool bSaveValue);
+  size_t CountChoiceListItems(bool bSaveValue);
   absl::optional<WideString> GetChoiceListItem(int32_t nIndex, bool bSaveValue);
   bool IsChoiceListMultiSelect();
   bool IsChoiceListCommitOnSelect();
